add const to some members of MetisGraphPartitioner. Put CoarsenMetisMaterialID class into its own file
